Refactor add_proto_file helper
Fix tmp file structure
